Move creatable macro types to /models
