+++ src/functions/macros/get-macros-authorized-to-me.spec.ts
@@ -7,12 +7,12 @@
  **************************************************************************/
 
 import { random, sortBy } from 'lodash';
-import { isMacro, Macro, User } from '../../models';
+import { CreatableMacro, isMacro, Macro, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { CreatableUser, makeCreateOneUser, makeGetOneUser } from '../users';
-import { CreatableMacro, makeCreateOneMacro } from './create-one-macro';
+import { makeCreateOneMacro } from './create-one-macro';
 import { makeDeleteOneMacro } from './delete-one-macro';
 import { makeGetAllMacros } from './get-all-macros';
 import { makeGetMacrosAuthorizedToMe } from './get-macros-authorized-to-me';
