+++ src/functions/macros/get-macros-by-group.spec.ts
@@ -7,14 +7,14 @@
  **************************************************************************/
 
 import { random } from 'lodash';
-import { CreatableGroup, isMacro, User } from '../../models';
+import { CreatableGroup, CreatableMacro, isMacro, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { makeAddOneUserToManyGroups } from '../groups/add-one-user-to-many-groups';
 import { makeCreateOneGroup } from '../groups/create-one-group';
 import { CreatableUser, makeCreateOneUser, makeGetMyUser, makeGetOneUser } from '../users';
-import { CreatableMacro, makeCreateOneMacro } from './create-one-macro';
+import { makeCreateOneMacro } from './create-one-macro';
 import { makeDeleteOneMacro } from './delete-one-macro';
 import { makeGetAllMacros } from './get-all-macros';
 import { makeGetMacrosByGroup } from './get-macros-by-group';
