+++ src/functions/macros/get-macros-by-user.spec.ts
@@ -7,12 +7,12 @@
  **************************************************************************/
 
 import { random } from 'lodash';
-import { isMacro, User } from '../../models';
+import { CreatableMacro, isMacro, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { CreatableUser, makeCreateOneUser, makeGetOneUser } from '../users';
-import { CreatableMacro, makeCreateOneMacro } from './create-one-macro';
+import { makeCreateOneMacro } from './create-one-macro';
 import { makeDeleteOneMacro } from './delete-one-macro';
 import { makeGetAllMacros } from './get-all-macros';
 import { makeGetMacrosByUser } from './get-macros-by-user';
