+++ src/models/macro/creatable-macro.ts
@@ -6,10 +6,17 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-macro';
-export * from './is-macro';
-export * from './macro';
-export * from './raw-creatable-macro';
-export * from './raw-macro';
-export * from './to-macro';
-export * from './to-raw-creatable-macro';
+import { NumericID } from '../../value-objects';
+
+export interface CreatableMacro {
+	groupIDs?: Array<NumericID>;
+
+	/**
+	 * All uppercase and no spaces.
+	 */
+	name: string;
+	description?: string | null;
+	labels?: Array<string>;
+
+	expansion: string;
+}
