+++ src/models/macro/raw-creatable-macro.ts
@@ -6,10 +6,12 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-export * from './creatable-macro';
-export * from './is-macro';
-export * from './macro';
-export * from './raw-creatable-macro';
-export * from './raw-macro';
-export * from './to-macro';
-export * from './to-raw-creatable-macro';
+import { RawNumericID } from '../../value-objects';
+
+export interface RawCreatableMacro {
+	GIDs: Array<RawNumericID>;
+	Name: string;
+	Description: string | null;
+	Expansion: string;
+	Labels: Array<string>;
+}
